Add new Specifications container classes
Adds a new Specifications class, and specific NamespaceSpecifications and DataElementSpecifications classes.  These server as a good container for definitions, with functions for adding and looking up definitions.  This should be the main way specifications are passed around between tools (for example, the importer will return an instance of Specifications).

Also added grammarVersion to the DataElement class so that tools no what version of the grammar was used to define the data element.