Add getPossibleIdentifiers
The current possibleIdentifiers getter also returns the types from includesType constraints.  We don't always want that.  I've added a new function that takes an argument to determine whether or on to use includesType identifiers.

Now there is overlap between the function and the getter.  This should be resolved, but currently would impact other modules.